Updated Sparkle, building 64bit now.

+++ CCMenu/Classes/CCMServerMonitor.m
@@ -89,7 +89,7 @@ NSString *CCMProjectStatusUpdateNotification = @"CCMProjectStatusUpdateNotificat
 	if([[self connections] count] == 0)
 		return;
 	[NSTimer scheduledTimerWithTimeInterval:0.1 target:self selector:@selector(pollServers:) userInfo:nil repeats:NO];
-	int interval = [defaultsManager pollInterval];
+	NSInteger interval = [defaultsManager pollInterval];
 	timer = [NSTimer scheduledTimerWithTimeInterval:interval target:self selector:@selector(pollServers:) userInfo:nil repeats:YES];
 }
